Move TimeFlagsTransform to transforms/timestamp/time_flags.py
